Add PremiumCalculator utils
