+++ package.json
@@ -29,7 +29,7 @@
     "@types/chai": "^4.2.11",
     "@types/chai-as-promised": "7.1.0",
     "@types/mocha": "^5.2.7",
-    "@types/node": "^14.0.13",
+    "@types/node": "^16.11.7",
     "@types/sinon": "^9.0.4",
     "@typescript-eslint/eslint-plugin": "^4.18.0",
     "@typescript-eslint/parser": "^4.18.0",
