+++ packages/contract/package.json
@@ -25,7 +25,6 @@
   "dependencies": {
     "@atomic-utils/format": "^0.1.3",
     "@atomicfinance/bitcoin-networks": "^2.4.1",
-    "@atomicfinance/constants": "^0.1.0",
     "@liquality/bitcoin-networks": "^1.12.0",
     "@node-dlc/core": "^0.12.4",
     "@node-dlc/logger": "0.12.4",
